+++ Src/tablereader.cpp
@@ -85,7 +85,7 @@ void TableReader::ReadTable()
 
         for(size_t x = 0; x<elems.size(); ++x)
         {
-            storage->SetCell(x,y,ICell::CellFactureMethod(elems[x]));
+            storage->SetCell((int)x,y,ICell::CellFactureMethod(elems[x]));
         }
     }
     delete[] buffer;
